refactor(shared-types): declare the analytics history contract once (fleetsim-all-i19m)
The `/analytics/history` row, bucket, metadata and envelope shapes were
declared twice: once in the simulator's `StateStore.ts`, and once by hand in
the UI's `useAnalytics.ts`, because `shared-types` carried no analytics types
and the UI cannot import from the simulator. The UI's copy had already drifted
-- `AnalyticsBucketMeta.aggregation` was optional and typed
`Record<string, string>` where the server always sends it as a known const --
and the counter-vs-mean split was restated as a literal with a hand-written
test locking it against the simulator's map.

They now live in `packages/shared-types/src/rest.ts`, alongside the other REST
DTOs, for the reason that file already gives: a changed payload shape should
fail to compile on the consumer side rather than drift silently.

Three things get stronger rather than merely deduplicated:

- The route's `?envelope=true` body is now typed `AnalyticsHistoryEnvelope`
  instead of an inline object literal, so the body the simulator writes and the
  body the UI parses are checked against one declaration.
- The UI's `BUCKET_AGGREGATION` is DERIVED from
  `ANALYTICS_BUCKET_AGGREGATION` by indexing it at `summary.<measure>`, so
  renaming a measure in the contract fails to compile in the UI. Its
  hand-copied parity test is replaced by one asserting the narrowing is
  faithful.
- A new type test asserts every `AnalyticsSummary` and `FleetAnalytics` measure
  has a declared fold. Verified it bites: adding a field to `AnalyticsSummary`
  without an aggregation entry fails `tsc` with "summary.avgDwellTime is not
  assignable".

`StateStore` re-exports the shared declarations so its existing importers keep
one import site, and keeps what is genuinely store-local: `AnalyticsHistoryOptions`
(how a query is asked for) and `AnalyticsHistoryResult` (the array-with-hidden-meta
it hands back in-process, which never crosses the wire).
